Fixed bugs

+++ src/Core/CityBookMVCOnionApplication/ViewModels/Review/IncludeReviewVM.cs
@@ -7,7 +7,7 @@ namespace CityBookMVCOnionApplication.ViewModels
         public int Id { get; init; }
         public string Name { get; init; }
         public string? Text { get; init; }
-        public float RatingStar { get; init; }
+        public int RatingStar { get; init; }
         public string UserId { get; init; }
         public IncludeUserVM User { get; init; }
     }
