+++ src/Core/CityBookMVCOnionApplication/ViewModels/Review/ItemReviewVM.cs
@@ -5,7 +5,7 @@
         public int Id { get; init; }
         public string? Text { get; init; }
         public string Name { get; init; }
-        public float RatingStar { get; init; }
+        public int RatingStar { get; init; }
 
         public int UserId { get; init; }
         public IncludeUserVM User { get; init; }
