+++ src/Core/CityBookMVCOnionApplication/ViewModels/Review/UpdateReviewVM.cs
@@ -4,7 +4,7 @@
     {
         public string? Text { get; init; }
         public string Name { get; init; }
-        public float RatingStar { get; init; }
+        public int RatingStar { get; init; }
 
         public int UserId { get; init; }
         public int PlaceId { get; init; }
